Move helpers out of main module
